feat(deploy): add automated script for deploy
This change adds an automated script for deploying the infrastructure
in this sample.

In the root directory of this project, run `npm install` then execute
the deploy script by running `./deploy.js`

Documentation has been added to the README.md in docs/en as well as
docs/ja, however, the Japanese version will likely need edits from
a native speaker.

A `deploy-settings.json` file gets created when a user runs the
deploy script. This settings file is used to pass variables between
stages and is also used to skip steps that have been completed
previously in the event of needing to run the script more than once.

Closes #5